This might fix the broken list of recent file on Linux.
I can't be sure, because I can't reproduce the problem, although I'm currently writing this in my Ubuntu VM.  Here, it works flawlessly. Maybe, a corrupt QSettings file is the cause...

On another note: I added the precompiled _parkin.so (and parkin.py because this was missing, too). Let's see if it makes sense to include these precompiled PARKINcpp binaries...