Show data source in legend when plotting same-entity data from different sources.
This also fixes the issue in the case of the sensitivity overview. Now, when e.g. Species1 is selected for more than on Parameter, then the legend shows "Species1 (ParamA)" and "Species1 (ParamB)".

But, this is also useful when plotting data from the same e.g. Species1 coming from different files.

fixes #7